Revert "Extract time parsing logic Move time parsing logic to a separate module to separate concerns"
This reverts commit ffdd1a9b6abb9009b5679fe8d57cdc331e6b5d61.
